+++ android/src/main/kotlin/tv/mta/flutter_playout/video/PlayerLayout.java
@@ -15,6 +15,7 @@ import android.os.IBinder;
 import android.support.v4.media.MediaMetadataCompat;
 import android.support.v4.media.session.MediaSessionCompat;
 import android.support.v4.media.session.PlaybackStateCompat;
+import android.util.Log;
 import android.view.KeyEvent;
 
 import androidx.annotation.NonNull;
@@ -45,6 +46,7 @@ import tv.mta.flutter_playout.PlayerState;
 import tv.mta.flutter_playout.R;
 
 public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventChannel.StreamHandler {
+    private final String TAG = "PlayerLayout";
 
     public static SimpleExoPlayer activePlayer;
 
@@ -468,9 +470,13 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                 message.put("offset", eventTime.currentPlaybackPositionMs / 1000);
 
+                Log.d(TAG, "onSeek: [position=" + beforeSeek + "] [offset=" +
+                        eventTime.currentPlaybackPositionMs / 1000 + "]");
                 eventSink.success(message);
 
-            } catch (Exception e) { /* ignore */ }
+            } catch (Exception e) {
+                Log.e(TAG, "onSeek: ", e);
+            }
         }
 
         @Override
@@ -493,9 +499,12 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                 message.put("error", errorMessage);
 
+                Log.d(TAG, "onError: [errorMessage=" + errorMessage + "]");
                 eventSink.success(message);
 
-            } catch (Exception e) { /* ignore */ }
+            } catch (Exception e) {
+                Log.e(TAG, "onError: ", e);
+            }
         }
 
         @Override
@@ -513,9 +522,12 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                         message.put("name", "onPlay");
 
+                        Log.d(TAG, "onPlay: []");
                         eventSink.success(message);
 
-                    } catch (Exception e) { /* ignore */ }
+                    } catch (Exception e) {
+                        Log.e(TAG, "onPlay: ", e);
+                    }
 
                 } else {
 
@@ -527,9 +539,12 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                         message.put("name", "onPause");
 
+                        Log.d(TAG, "onPause: []");
                         eventSink.success(message);
 
-                    } catch (Exception e) { /* ignore */ }
+                    } catch (Exception e) {
+                        Log.e(TAG, "onPause: ", e);
+                    }
 
                 }
 
@@ -545,9 +560,12 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                     message.put("name", "onComplete");
 
+                    Log.d(TAG, "onComplete: []");
                     eventSink.success(message);
 
-                } catch (Exception e) { /* ignore */ }
+                } catch (Exception e) {
+                    Log.e(TAG, "onComplete: ", e);
+                }
 
             }
         }
@@ -572,10 +590,13 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                         message.put("time", mPlayerView.getCurrentPosition() / 1000);
 
+                        Log.d(TAG, "onTime: [time=" + mPlayerView.getCurrentPosition() / 1000 + "]");
                         eventSink.success(message);
                     }
 
-                } catch (Exception e) { /* ignore */ }
+                } catch (Exception e) {
+                    Log.e(TAG, "onTime: ", e);
+                }
 
                 onDuration();
 
@@ -625,7 +646,7 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
             long newDuration = mPlayerView.getDuration();
 
-            if (newDuration != mediaDuration) {
+            if (newDuration != mediaDuration && eventSink != null) {
 
                 mediaDuration = newDuration;
 
@@ -635,10 +656,13 @@ public class PlayerLayout extends PlayerView implements FlutterAVPlayer, EventCh
 
                 message.put("duration", mediaDuration);
 
+                Log.d(TAG, "onDuration: [duration=" + mediaDuration + "]");
                 eventSink.success(message);
             }
 
-        } catch (Exception e) { /* ignore */ System.out.println(e); }
+        } catch (Exception e) {
+            Log.e(TAG, "onDuration: " + e.getMessage(), e);
+        }
     }
 
     @Override